Add back code that disappered Git, why you did it to me?!

+++ src/main/java/com/github/haseoo/courier/models/AddressModel.java
@@ -1,13 +1,10 @@
 package com.github.haseoo.courier.models;
 
-import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import org.mapstruct.Mapper;
 
 import javax.persistence.*;
-
-import java.util.List;
 
 import static javax.persistence.GenerationType.IDENTITY;
 
@@ -20,7 +17,7 @@ public class AddressModel {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     @Column(insertable = false, nullable = false)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false)
     private String city;
